Introduce home page header widget, update search widget and pass as homepage header widget
