+++ app/src/main/kotlin/uk/govuk/app/ui/GovUkApp.kt
@@ -317,6 +317,17 @@ private fun GovUkNavHost(
                 onSuppressClick = onSuppressWidgetClick
             ),
             modifier = Modifier.padding(paddingValues),
+            headerWidget = if (homeWidgets?.contains(HomeWidget.SEARCH) == true) {
+                { modifier ->
+                    SearchWidget(
+                        onClick = { text ->
+                            onWidgetClick(text, false)
+                            navController.navigate(SEARCH_GRAPH_ROUTE)
+                        },
+                        modifier = modifier
+                    )
+                }
+            } else null,
             transitionOverrideRoutes = listOf(SEARCH_GRAPH_ROUTE)
         )
         settingsGraph(
@@ -370,19 +381,6 @@ private fun homeScreenWidgets(
                 }
             }
 
-            HomeWidget.SEARCH -> {
-                widgets.add { modifier ->
-                    SearchWidget(
-                        onClick = { text ->
-                            onClick(text, false)
-                            navController.navigate(SEARCH_GRAPH_ROUTE)
-                        },
-                        modifier = modifier
-                    )
-                    LargeVerticalSpacer()
-                }
-            }
-
             HomeWidget.RECENT_ACTIVITY -> {
                 widgets.add { modifier ->
                     VisitedWidget(
@@ -416,6 +414,8 @@ private fun homeScreenWidgets(
                     LargeVerticalSpacer()
                 }
             }
+
+            else -> { } // Do nothing
         }
     }
     return widgets
